// src/api/generated/schema.d.ts
export interface components {
    schemas: {
// ...
        GoalDetail: components["schemas"]["GoalBase"] & {
            /** Format: double */
            current_value: number;
            /** Format: double */
            remaining_value: number;
            /** Format: double */
            percent_complete: number;
            days_left: number;
            days_left_for_pace: number;
            days_total: number;
            days_elapsed: number;
            /** Format: double */
            pace_expected_per_day: number;
            /** Format: double */
            pace_required_per_day: number;
            /** Format: double */
            pace_current_7d: number;
            /** Format: double */
            pace_current_30d: number;
            /** Format: double */
            pace_current_all: number;
            /** Format: date */
            eta_date: string | null;
            /** Format: double */
            expected_by_today: number;
            /** Format: double */
            behind_value: number;
            /** Format: double */
            catchup_pace_next_7_days: number;
        };
// ...
    };
// ...
}